Split history feature module tests
Replace the broad history feature module spec with behavior-owned specs and shared harness support while leaving deterministic render assertions with presenter/view tests.
