+++ apps/ui/tests/history_feature_navigation.spec.ts
@@ -1,0 +1,17 @@
+import { expect, test } from "vitest";
+import {
+  createFeatureHarness,
+  installHistoryFeatureTestLifecycle,
+} from "./history_feature_test_support";
+
+installHistoryFeatureTestLifecycle();
+
+test("history feature routes the open-live action to dashboard navigation", () => {
+  const { feature, getLatestHandlers, primaryViewActivations } =
+    createFeatureHarness();
+  feature.bindHandlers();
+
+  getLatestHandlers()?.onTableInteraction({ type: "open-live" });
+
+  expect(primaryViewActivations).toEqual(["dashboardView"]);
+});
